Very Shitty Explorer. I have no idea if this works. We need to integrate our code to figure that out. Also path finding based on landmarks is not implemented yet.

+++ src/main/java/Group5/Agent/Edge.java
@@ -4,10 +4,10 @@ public class Edge implements Comparable<Edge> {
     private final String id;
     private final Vertex source;
     private final Vertex destination;
-    private final int weight;
+    private final double weight;
     private Edge mirror;
 
-    public Edge(String id, Vertex source, Vertex destination, int weight) {
+    public Edge(String id, Vertex source, Vertex destination, double weight) {
         this.id = id;
         this.source = source;
         this.destination = destination;
@@ -34,7 +34,7 @@ public class Edge implements Comparable<Edge> {
         return source;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
